+++ functions/vndaProductListingPage.ts
@@ -2,8 +2,8 @@ import type { LiveState } from "$live/types.ts";
 import type { LoaderFunction } from "$live/types.ts";
 import { createClient } from "../commerce/vnda/client.ts";
 import type { ProductListingPage } from "../commerce/types.ts";
+import { VNDA_SORT_OPTIONS } from "../commerce/vnda/client.ts";
 import { ConfigVNDA, VNDASort } from "../commerce/vnda/types.ts";
-import { VNDA_SORT_OPTIONS } from "../commerce/vnda/constants.ts";
 
 import {
   filtersFromSearchParams,
